FIX: 리포지토리 유저 참조 수정

+++ src/main/java/alpha/chupchup/service/RecipeService.java
@@ -24,15 +24,16 @@ public class RecipeService {
     private final UserRepository userRepository;
 
     @Transactional
-    public void postRecipeFavorite(Long recipeId) {
-        User user = userRepository.findById(0L)
-                .orElseThrow(() -> new RuntimeException("사용자를 찾을 수 없습니다."));
-        Recipe recipe = recipeRepository.findById(recipeId)
-                .orElseThrow(() -> new RuntimeException("레시피를 찾을 수 없습니다."));
-        Optional<UserRecipeFavorite> recipeFavorite = userRecipeFavoriteRepository.findByUserAndRecipe(user, recipe);
+    public void postRecipeFavorite(Long userId, Long recipeId) {
+        Optional<UserRecipeFavorite> recipeFavorite = userRecipeFavoriteRepository.findByUserIdAndRecipeId(userId, recipeId);
         if (recipeFavorite.isPresent()) {
             throw new RuntimeException("이미 등록된 좋아요입니다.");
         }
+
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new RuntimeException("사용자를 찾을 수 없습니다."));
+        Recipe recipe = recipeRepository.findById(recipeId)
+                .orElseThrow(() -> new RuntimeException("레시피를 찾을 수 없습니다."));
         UserRecipeFavorite favorite = UserRecipeFavorite.builder()
                 .user(user)
                 .recipe(recipe)
@@ -41,20 +42,18 @@ public class RecipeService {
     }
 
     @Transactional
-    public void deleteRecipeFavorite(Long recipeId) {
-        User user = userRepository.findById(0L)
-                .orElseThrow(() -> new RuntimeException("사용자를 찾을 수 없습니다."));
-        Recipe recipe = recipeRepository.findById(recipeId)
-                .orElseThrow(() -> new RuntimeException("레시피를 찾을 수 없습니다."));
-        UserRecipeFavorite favorite = userRecipeFavoriteRepository.findByUserAndRecipe(user, recipe)
+    public void deleteRecipeFavorite(Long userId, Long recipeId) {
+        UserRecipeFavorite favorite = userRecipeFavoriteRepository.findByUserIdAndRecipeId(userId, recipeId)
                 .orElseThrow(() -> new RuntimeException("즐겨찾기 기록이 존재하지 않습니다."));
-        userRecipeFavoriteRepository.delete(favorite);
+        if (favorite.getUser().getId().equals(userId)) {
+            userRecipeFavoriteRepository.delete(favorite);
+        } else {
+            throw new RuntimeException("해당 유저가 등록한 즐겨찾기가 아닙니다.");
+        }
     }
 
     public List<RecipeResponseDto> getFavoriteRecipe(Long userId) {
-        User user = userRepository.findById(userId)
-                        .orElseThrow(() -> new RuntimeException("사용자를 찾을 수 없습니다."));
-        List<UserRecipeFavorite> recipes = userRecipeFavoriteRepository.findAllByUser(user);
+        List<UserRecipeFavorite> recipes = userRecipeFavoriteRepository.findAllByUserId(userId);
 
         return recipes.stream()
                 .map(fav -> {
